NYCCHKBK-47 vendor_id changed to int from bigint in contracts aggregate table
git-svn-id: https://omb-dashboardplatform.reisys.com:10039/NYC_Checkbook_2_0/Database@1813 948744eb-f611-4f28-8820-250fb6d8f47d

Former-commit-id: a59cc29db3bc8b75690e9173efb11c3948e6a824

+++ checkbook/DB scripts/VerificationAndAggregateScripts/Contract_Aggregates.sql
@@ -302,7 +302,7 @@ CREATE TABLE aggregateon_contracts_cumulative_spending(
 	master_agreement_yn character(1),
 	description varchar,
 	contract_number varchar,
-	vendor_id bigint,
+	vendor_id int,
 	award_method_id smallint,
 	agency_id smallint,
 	original_contract_amount numeric(16,2),
